requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
scikit-learn>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
folium>=0.16.0
streamlit-folium>=0.18.0
joblib>=1.3.0
pytz>=2024.1
